Types: Rename to types.ts, simplify & fix typing issues
