+++ src/Player/playerMovementUtils.ts
@@ -1,20 +1,20 @@
-import { PlayerPosition, PlayerVelocity } from '../allTypes';
+import { PlayerPosition, PlayerVelocity } from '../types';
 /**
  * serialize is a helper function that is a light wrapper around stringifying a JSON object.
- * @param data 
+ * @param data
  * @returns string
  */
-export function serialize(data:PlayerPosition):string {
+export function serialize(data: PlayerPosition): string {
   return JSON.stringify(data);
 }
 
 /**
  * deserialize is a helper function that is a light wrapper around parsing a JSON string.
  * Returns undefined if error occured.
- * @param data 
+ * @param data
  * @returns string
  */
-export function deserialize(data:PlayerPosition) {
+export function deserialize(data: PlayerPosition) {
   try {
     return JSON.parse(data.toString());
   } catch (e) {
@@ -23,24 +23,26 @@ export function deserialize(data:PlayerPosition) {
 }
 
 /**
- * round is a simple helper funciton that rounds a float 
+ * round is a simple helper funciton that rounds a float
  * to a number of places.
- * @param float 
- * @param numToRound 
+ * @param float
+ * @param numToRound
  * @returns rounded number
  */
-export function round(float:Number, numToRound:number):number{
+export function round(float: Number, numToRound: number): number {
   return parseFloat(float.toFixed(numToRound));
 }
 /**
  * roundEntriesInVector is a helper function that rounds a vector
  * to a number of places.
- * @param vector 
- * @param numToRound 
+ * @param vector
+ * @param numToRound
  * @returns vector
  */
-export function roundEntriesInVector(vector:PlayerPosition | PlayerVelocity, numToRound:number){
-
+export function roundEntriesInVector(
+  vector: PlayerPosition | PlayerVelocity,
+  numToRound: number
+) {
   for (let [key, value] of Object.entries(vector)) {
     vector[key as keyof PlayerPosition] = round(value, numToRound);
   }
