--- src/allTypes.ts
@@ -1,11 +1,0 @@
-export type PlayerPosition = {
-  x: number;
-  y: number;
-  z: number;
-};
-
-export type PlayerVelocity = {
-  x: number;
-  y: number;
-  z: number;
-};
